historiaclinica/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
py==1.10.0
PyJWT==2.1.0
SQLAlchemy==1.4.15